refactor(youtube): one owner for audible playback (refs #628)
The autoplay-policy knowledge under the play-then-pause bug class lived
in comments smeared across bridge, session, and events, and its timing
only held jointly: 3x250ms pause-confirm < 900ms heal < 1200ms hint — a
three-way invariant spanning three files with no owner. de29a5be called
its predecessor "structurally doomed" for exactly this reason: it fixed
the route while the muted-start -> restore -> heal cycle stayed split.

New YoutubeAudiblePlaybackPolicy owns the whole choreography and its
clock:

- onPlaying arms the per-document progress-gated restore (once per
  document — redundant unMutes are pause triggers under the gesture
  lock); onPlaybackProgress gates on advancing samples plus the platform
  settle delay; a fallback timer restores when progress never arrives;
  _restoreVolume re-applies volume, pins the document, and schedules the
  one-shot post-restore heal; onPause/cancelPending abandon pending
  restores.
- Timing constants move in: windowsVolumeRestoreDelay (400ms),
  volumeRestoreFallback (1500ms, was on the session), and
  postRestoreHealDelay (900ms), plus pollTick (250ms) as the canonical
  poll cadence — the poll loop now ticks on the policy's constant.
- The joint invariant is asserted as a test (confirm window < heal <
  hint), so the constants can no longer drift independently.
- YoutubeWebViewEvents shrinks to pure dispatch: transport transitions to
  the session (per #627), audibility to the policy. It loses its volume
  ctor params, timers, and five private helpers.
- The session keeps the latches (state store); its tapToPlayHintDelay is
  now public and referenced by the invariant test.

Volume/heal behaviour tests move from the engine test file to
youtube_session_audible_policy_test.dart, driving the policy directly.
