+++ lib/features/player/application/engines/youtube/youtube_webview_events.dart
@@ -1,13 +1,15 @@
 /// DOM `<video>` event dispatch for [YoutubeWebViewController].
+///
+/// Dispatch only: transport transitions go to [YoutubeSession], the
+/// audible-start choreography to [YoutubeAudiblePlaybackPolicy] (issue #628).
 library;
 
 import 'dart:async';
 
-import 'package:flutter/foundation.dart'
-    show defaultTargetPlatform, TargetPlatform;
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 
 import 'package:enjoy_player/core/logging/log.dart';
+import 'package:enjoy_player/features/player/application/engines/youtube/youtube_audible_playback_policy.dart';
 import 'package:enjoy_player/features/player/application/engines/youtube/youtube_session.dart';
 import 'package:enjoy_player/features/player/application/engines/youtube/youtube_video_event.dart';
 import 'package:enjoy_player/features/player/application/engines/youtube/youtube_webview_bridge.dart';
@@ -18,11 +20,6 @@ typedef YoutubeSeekFn = Future<void> Function(Duration target);
 typedef YoutubePollStartFn = void Function();
 typedef YoutubePollStopFn = void Function();
 typedef YoutubeFirstPlayingFn = void Function();
-typedef YoutubeReapplyVolumeFn = Future<void> Function();
-
-/// Injectable post-restore heal play (defaults to
-/// [YoutubeWebViewBridge.play], which preserves audible state).
-typedef YoutubeHealPlayFn = Future<void> Function(InAppWebViewController? web);
 
 /// Handles `onVideoEvent` JavaScript callbacks from the watch page.
 class YoutubeWebViewEvents {
@@ -32,53 +29,19 @@ class YoutubeWebViewEvents {
     required this.onFirstPlaying,
     required this.startPolling,
     required this.stopPolling,
-    required this.reapplyVolume,
     required this.seekTo,
-    Duration? volumeRestoreDelay,
-    Duration? volumeRestoreFallback,
-    Duration? postRestoreHealDelay,
-    YoutubeHealPlayFn? healPlay,
-  }) : volumeRestoreDelay =
-           volumeRestoreDelay ??
-           (defaultTargetPlatform == TargetPlatform.windows
-               ? windowsVolumeRestoreDelay
-               : Duration.zero),
-       volumeRestoreFallback =
-           volumeRestoreFallback ?? YoutubeSession.volumeRestoreFallback,
-       postRestoreHealDelay =
-           postRestoreHealDelay ??
-           YoutubeWebViewEvents.defaultPostRestoreHealDelay,
-       healPlay = healPlay ?? YoutubeWebViewBridge.play;
-
-  /// Legacy minimum settle before progress-gated unmute (Windows first play).
-  /// Progress confirmation is preferred; this delay is only a lower bound when
-  /// progress arrives earlier than the platform settle window.
-  static const Duration windowsVolumeRestoreDelay = Duration(milliseconds: 400);
-
-  /// How long after a volume restore to check whether the unmute tripped the
-  /// WebView's autoplay gesture lock (which pauses the element). Must exceed
-  /// the poll loop's pause-confirmation window (~3 × 250 ms) so a real pause
-  /// is already reflected in [YoutubeSession.playing]; must stay short enough
-  /// to beat the tap-to-play recovery hint.
-  static const Duration defaultPostRestoreHealDelay = Duration(
-    milliseconds: 900,
-  );
-
-  final Duration volumeRestoreDelay;
-  final Duration volumeRestoreFallback;
-  final Duration postRestoreHealDelay;
+    required this.audibility,
+  });
 
   final YoutubeSession session;
   final InAppWebViewController? Function() webController;
   final YoutubeFirstPlayingFn onFirstPlaying;
   final YoutubePollStartFn startPolling;
   final YoutubePollStopFn stopPolling;
-  final YoutubeReapplyVolumeFn reapplyVolume;
   final YoutubeSeekFn seekTo;
-  final YoutubeHealPlayFn healPlay;
 
-  Timer? _volumeRestoreFallbackTimer;
-  DateTime? _volumeRestoreArmedAt;
+  /// Audible-start choreography owner (mute-start → restore → heal).
+  final YoutubeAudiblePlaybackPolicy audibility;
 
   dynamic handle(List<dynamic> args) {
     if (args.isEmpty) return null;
@@ -90,32 +53,24 @@ class YoutubeWebViewEvents {
         _logEvents.fine('youtube video play requested vid=${session.videoId}');
         break;
       case YoutubeVideoEventName.playing:
-        // A fresh watch document (cold open or post-ad reload) starts muted;
-        // arm the progress-gated restore once. Later `playing` events in an
-        // already-restored document must NOT re-unmute: every programmatic
-        // unMute is a pause trigger under Chromium's autoplay gesture lock
-        // (the play-then-pause root cause), so redundancy here is the bug.
-        final needsRestore = session.needsVolumeRestore;
         session.notePlayingConfirmed();
         onFirstPlaying();
         session.emitBuffering(false);
         startPolling();
         applyPendingSeek();
-        if (needsRestore) {
-          _armProgressGatedVolumeRestore();
-        }
+        audibility.onPlaying();
         _logEvents.fine('youtube video playing vid=${session.videoId}');
         break;
       case YoutubeVideoEventName.pause:
-        // Do NOT reset [pausedPollStreak] — a DOM pause while Dart still thinks
+        // Do NOT reset the pause streak — a DOM pause while Dart still thinks
         // playing must accumulate toward poll confirmation. Resetting here
         // previously extended the stale-`playing` window and made the next
         // transport toggle issue `pause()` instead of `play()`.
-        cancelPendingVolumeRestore();
+        audibility.onPause();
         _logEvents.fine('youtube video paused vid=${session.videoId}');
         break;
       case YoutubeVideoEventName.playRejected:
-        cancelPendingVolumeRestore();
+        audibility.cancelPending();
         session.noteUserPlayUnresolved();
         final reason = args.length > 1 ? '${args[1]}' : 'unknown';
         _logEvents.warning(
@@ -127,7 +82,7 @@ class YoutubeWebViewEvents {
         session.scheduleRecoveryHint();
         break;
       case YoutubeVideoEventName.ended:
-        cancelPendingVolumeRestore();
+        audibility.cancelPending();
         session.noteEnded();
         stopPolling();
         unawaited(YoutubeWebViewBridge.pauseVideoElement(webController()));
@@ -151,7 +106,7 @@ class YoutubeWebViewEvents {
         }
         break;
       case YoutubeVideoEventName.error:
-        cancelPendingVolumeRestore();
+        audibility.cancelPending();
         _logEvents.warning('YouTube video element error');
         session.noteUserPlayUnresolved();
         break;
@@ -163,87 +118,6 @@ class YoutubeWebViewEvents {
         break;
     }
     return null;
-  }
-
-  /// Called from the poll loop when `<video>.currentTime` advances.
-  void onPlaybackProgress(Duration position) {
-    if (!session.volumeRestorePending || session.disposed) return;
-    if (!session.playing) return;
-    final armedAt = _volumeRestoreArmedAt;
-    if (armedAt != null) {
-      final elapsed = DateTime.now().difference(armedAt);
-      if (elapsed < volumeRestoreDelay) return;
-    }
-    if (session.noteProgressForVolumeRestore(position)) {
-      unawaited(_restoreVolume(reason: 'progress'));
-    }
-  }
-
-  void _armProgressGatedVolumeRestore() {
-    cancelPendingVolumeRestore();
-    session.armVolumeRestorePending(baseline: session.lastPosition);
-    _volumeRestoreArmedAt = DateTime.now();
-    _logEvents.fine(
-      'youtube volume restore armed vid=${session.videoId} '
-      'fallbackMs=${volumeRestoreFallback.inMilliseconds} '
-      'minDelayMs=${volumeRestoreDelay.inMilliseconds}',
-    );
-    _volumeRestoreFallbackTimer = Timer(volumeRestoreFallback, () {
-      _volumeRestoreFallbackTimer = null;
-      if (session.disposed ||
-          !session.playing ||
-          !session.volumeRestorePending) {
-        return;
-      }
-      unawaited(_restoreVolume(reason: 'fallback'));
-    });
-  }
-
-  Future<void> _restoreVolume({required String reason}) async {
-    if (session.disposed || !session.playing) return;
-    session.clearVolumeRestorePending();
-    _volumeRestoreFallbackTimer?.cancel();
-    _volumeRestoreFallbackTimer = null;
-    _volumeRestoreArmedAt = null;
-    try {
-      await reapplyVolume();
-      session.noteVolumeRestored();
-      _logEvents.fine(
-        'youtube volume restored vid=${session.videoId} reason=$reason',
-      );
-      unawaited(_healPostRestorePause());
-    } on Object catch (error, stackTrace) {
-      _logEvents.warning(
-        'youtube volume restore failed vid=${session.videoId} reason=$reason',
-        error,
-        stackTrace,
-      );
-    }
-  }
-
-  /// One-shot self-heal for the gesture-lock failure mode: the unmute made
-  /// the element audible without user activation and the WebView paused it
-  /// right after ("Unmuting failed and the element was paused instead").
-  /// Re-issues play exactly once with audible state preserved — on engines
-  /// that allow audible starts this recovers invisibly; where the start is
-  /// genuinely refused, the poll loop's recovery hint remains the fallback.
-  Future<void> _healPostRestorePause() async {
-    await Future<void>.delayed(postRestoreHealDelay);
-    if (session.disposed ||
-        session.playing ||
-        session.playbackCompleted ||
-        session.needsVolumeRestore) {
-      return;
-    }
-    _logEvents.info('youtube post-restore pause heal vid=${session.videoId}');
-    unawaited(healPlay(webController()));
-  }
-
-  void cancelPendingVolumeRestore() {
-    _volumeRestoreFallbackTimer?.cancel();
-    _volumeRestoreFallbackTimer = null;
-    _volumeRestoreArmedAt = null;
-    session.clearVolumeRestorePending();
   }
 
   void applyPendingSeek() {
